+++ packages/hydrogen/src/WeaverseHydrogenRoot.tsx
@@ -5,6 +5,7 @@ import {
   WeaverseItemStore,
   WeaverseRoot,
   isBrowser,
+  useSafeExternalStore,
 } from '@weaverse/react'
 import {
   type ComponentType,
@@ -12,7 +13,6 @@ import {
   Suspense,
   createContext,
   memo,
-  useSyncExternalStore,
 } from 'react'
 import { ErrorBoundary } from 'react-error-boundary'
 import { defaultComponents } from '~/components'
@@ -39,6 +39,10 @@ export class WeaverseHydrogenItem extends WeaverseItemStore {
   constructor(initialData: HydrogenComponentData, weaverse: WeaverseHydrogen) {
     super(initialData, weaverse)
     let { data, ...rest } = initialData
+    if (!this.Element?.schema) {
+      console.error('Element is missing schema or not found!')
+      return
+    }
     let schemaData = generateDataFromSchema(this.Element.schema)
     Object.assign(this._store, schemaData, data, rest)
   }
@@ -147,7 +151,7 @@ export let WeaverseHydrogenRoot = memo(
     if (data) {
       if (data instanceof Promise) {
         return (
-          <ErrorBoundary fallbackRender={ErrorComponent}>
+          <ErrorBoundary fallbackRender={ErrorComponent as any}>
             <Suspense>
               <Await resolve={data}>
                 {(resolvedData) => (
@@ -184,7 +188,7 @@ export function withWeaverse(Component: ComponentType<any>) {
 
 export function useThemeSettings<T = HydrogenThemeSettings>() {
   let themeSettingsStore = useThemeSettingsStore()
-  let settings = useSyncExternalStore(
+  let settings = useSafeExternalStore(
     themeSettingsStore.subscribe,
     themeSettingsStore.getSnapshot,
     themeSettingsStore.getServerSnapshot,