+++ packages/shopify/src/index.ts
@@ -20,24 +20,25 @@ export * from './types'
 export * from './utils/fetch-project-data'
 
 export let registerThirdPartyElements = () => {
-  let elements = WeaverseShopify.integrations?.flatMap(
-    ({ elements }) => elements,
-  )
-  for (let { type, extraData } of elements) {
-    WeaverseShopify.registerElement({
-      type,
-      extraData,
-      Component: ThirdPartyElement.default,
-      defaultCss: ThirdPartyElement.css,
+  // biome-ignore lint/complexity/noForEach: <explanation>
+  WeaverseShopify.integrations
+    ?.flatMap(({ elements }) => elements)
+    .forEach(({ type, extraData }) => {
+      WeaverseShopify.registerElement({
+        type,
+        extraData,
+        Component: ThirdPartyElement.default,
+        defaultCss: ThirdPartyElement.css,
+      })
     })
-  }
 }
 
 export let registerShopifyElements = () => {
-  for (let elm of Object.values(SHOPIFY_ELEMENTS)) {
+  // biome-ignore lint/complexity/noForEach: <explanation>
+  Object.values(SHOPIFY_ELEMENTS).forEach((elm) => {
     // @ts-ignore
     WeaverseShopify.registerElement(elm)
-  }
+  })
   registerThirdPartyElements()
 }
 export class WeaverseShopify extends Weaverse {
@@ -65,6 +66,7 @@ export class WeaverseShopifyItem extends WeaverseItemStore {
 
   constructor(initialData: ElementData, weaverse: WeaverseShopify) {
     super(initialData, weaverse)
+    // @ts-expect-error
     let defaultData = this.Element?.Component?.defaultProps || {}
     let extraData = this.Element?.extraData
     Object.assign(this._store, defaultData, extraData, initialData)